Pronouns
Important: changed "his or her" to "their" to include non-binary people
Suggested: given Steve is changing their name to Brittany, there's a chance they *might* be trans, so "his" might be incorrect. Also changed to "their" given this ambiguity. I know this is a hypothetical person, but references to changing names and still using the pronoun typically associated with the previous name make me think of misgendering